feat: extract tags routes to routes/tags.py
