+++ templates/settings_help.html
@@ -80,7 +80,7 @@
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-info w-100"
-                       href="{{ url_for('view_tag_rules') }}">Tag Rules</a>
+                       href="{{ url_for('tags.view_tag_rules') }}">Tag Rules</a>
                 </div>
                 <div class="col-md-6">
                     <a class="btn btn-sm btn-outline-danger w-100"
